fix: 修复用户贡献提交联调和MyBatis启动配置

+++ src/main/java/com/camplus/login/entity/User.java
@@ -60,11 +60,11 @@ public class User {
     }
 
     // Getter 和 Setter
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
